+++ ydb/core/tx/replication/controller/secret_resolver.cpp
@@ -3,7 +3,7 @@
 #include "secret_resolver.h"
 
 #include <ydb/core/kqp/common/events/script_executions.h>
-#include <ydb/services/secret/resolver.h>
+#include <ydb/services/scheme_secret/resolver.h>
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/library/actors/core/actor_bootstrapped.h>
 #include <ydb/library/actors/core/hfunc.h>
