+++ ydb/core/tx/tiering/manager.cpp
@@ -2,7 +2,7 @@
 #include "manager.h"
 
 #include <ydb/core/base/appdata.h>
-#include <ydb/services/secret/resolver.h>
+#include <ydb/services/scheme_secret/resolver.h>
 #include <ydb/core/protos/config.pb.h>
 #include <ydb/core/tx/columnshard/columnshard_private_events.h>
 #include <ydb/core/tx/tiering/fetcher.h>
