+++ ydb/core/tx/tiering/ya.make
@@ -25,7 +25,7 @@ PEERDIR(
     ydb/public/api/protos
     ydb/services/bg_tasks/abstract
     ydb/services/metadata
-    ydb/services/secret
+    ydb/services/scheme_secret
 )
 
 YQL_LAST_ABI_VERSION()
