+++ ydb/services/scheme_secret/ut/common/ya.make
@@ -9,7 +9,7 @@ PEERDIR(
     ydb/library/aclib
     ydb/core/kqp/common/events
     ydb/core/kqp/ut/common
-    ydb/services/secret
+    ydb/services/scheme_secret
 )
 
 YQL_LAST_ABI_VERSION()
