+++ ydb/services/ya.make
@@ -22,7 +22,7 @@ RECURSE(
     persqueue_v1
     rate_limiter
     replication
-    secret
+    scheme_secret
     sqs_topic
     tablet
     test_shard
